+++ src/main/java/com/prgm/aroundthetown/room/controller/RoomController.java
@@ -1,7 +1,7 @@
 package com.prgm.aroundthetown.room.controller;
 
 import com.prgm.aroundthetown.room.dto.RequestCreateRoomDto;
-import com.prgm.aroundthetown.room.service.RoomService;
+import com.prgm.aroundthetown.room.service.RoomServiceImpl;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/v1")
 public class RoomController {
 
-    private RoomService roomService;
+    private RoomServiceImpl roomService;
 
     @PostMapping("/accommodation/{productId}")
     public void saveRoom(
